pyproject.toml
[tool.poetry]
name = "gen-code"
version = "0.1.0"
description = "repo for gen ai hack for everything code"
authors = ["Ishant Aggarwal <iagg1946@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
vertexai = "^0.0.1"
pip = "^24.0"
install = "^1.3.5"
google-cloud-aiplatform = ">=1.38"
langchain_google_vertexai = ">=0.1.0"
langchain = "^0.1.11"
gitpython = "^3.1.42"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
